requirements.txt
bytewax==0.11.1
textblob==0.17.1
python-dotenv==0.20.0
requests==2.28.1